requirements.txt
asgiref==3.2.3
autopep8==1.4.4
Django==3.0
django-filter==2.2.0
djangorestframework==3.10.3
Markdown==3.1.1
numpy==1.17.4
opencv-python==4.1.2.30
pycodestyle==2.5.0
pytz==2019.3
sqlparse==0.3.0
